coprocessor layer is named based on coprocessor ID
removed optimistion reflection overlay

it didn't show all optimisations that are now done in the TIA emulation
and it was difficult to present the information meaningfully. plus, it's
information about the emulation itself and not about the emulated machine.

cleaned up lint errors

+++ hardware/memory/cartridge/harmony/arm7tdmi/mam.go
@@ -27,7 +27,7 @@ const (
 func (m *mam) write(addr uint32, val uint32) bool {
 	switch addr {
 	case MAMCR:
-		m.mamcr = uint32(val)
+		m.mamcr = val
 	default:
 		return false
 	}
